requirements.txt
accelerate==0.21.0
addict==2.4.0
av==12.3.0
azure_storage==0.37.0
beautifulsoup4==4.12.3
bitsandbytes==0.41.0
dashscope==1.20.14
data_processing==0.1.4
decord==0.6.0
deepspeed==0.12.2
diffusers==0.31.0
diskcache==5.6.3
docker_py==1.10.6
easyocr==1.7.1
einops==0.6.1
einops_exts==0.0.4
evaluate==0.4.3
face_detection==0.2.2
fastapi==0.103.2
fastchat==0.1.0
fastembed==0.4.2
fiftyone==1.0.2
fire==0.6.0
fitz==0.0.1.dev2
flaml==2.3.2
flash_attn==2.7.0.post2
google_cloud_aiplatform==1.73.0
gradio==3.35.2
huggingface_hub==0.23.4
ipdb==0.13.13
ipython==8.12.3
latex2sympy2==1.9.1
llama_index==0.12.2
llmlingua==0.2.2
markdownify==0.14.1
matplotlib==3.9.0
mistralai==1.2.3
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
omegaconf==2.3.0
openai==1.3.5
packaging==24.2
pandas==2.0.3
pathvalidate==3.2.1
pdfminer==20191125
peft==0.4.0
pgvector==0.3.6
Pillow==9.5.0
portalocker==2.10.0
protobuf==4.25.3
psycopg==3.2.3
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==1.10.8
pydantic_core==2.18.4
PyJWT==2.10.1
pypdf==5.1.0
pysubs2==1.7.3
python-dotenv==1.0.1
PyYAML==6.0.2
qdrant_client==1.12.1
ram==0.1
redis==5.2.0
regex==2024.5.15
replicate==0.27.0
Requests==2.32.3
rich==13.7.1
ring_flash_attn==0.1.1
rouge==1.0.1
safetensors==0.4.3
sat==0.7.0
scikit_learn==1.5.2
scipy==1.14.0
seaborn==0.13.2
sentence_transformers==3.3.1
setuptools==75.3.0
sglang==0.3.6.post3
shortuuid==1.0.13
spacy==3.8.2
sty==1.0.6
supervision==0.6.0
tabulate==0.9.0
termcolor==2.4.0
tiktoken==0.7.0
timm==0.6.13
tokenizers==0.19.1
torch==2.1.2
torch_npu==2.4.0
torchvision==0.16.2
tqdm==4.67.0
transformers==4.43.0
typer==0.12.3
typing_extensions==4.12.2
tyro==0.8.8
unstructured==0.16.9
uvicorn==0.30.1
validators==0.28.3
websocket_client==1.8.0
wikipedia==1.4.0
wolframalpha==5.1.3
xtuner==0.1.11
yapf==0.40.2
